requirements.txt
PyYAML==3.11
bottle==0.12.8
pyserial==2.7
wsgiref==0.1.2
